Changed openapi to generate kotlin server instead of java server spec
